Included file check

+++ plot.py
@@ -1,11 +1,12 @@
 import csv
+import os
 from matplotlib.pyplot import figure
 import numpy as np
 import matplotlib.pylab as plt
   
 
 numeri= [100, 1000, 5000, 20000, 50000, 1000000, 1000000];
-range = [10, 50, 100, 500];
+range = [10, 50, 100, 500, 1000, 10000, 50000, 100000];
 offset = [100, 500, 1000];
 
 for n in numeri:
@@ -16,33 +17,34 @@ for n in numeri:
       oReal = o
 
       fileName = "pi" + str(nReal) + "-" + str(rReal) + "-" + str(oReal)
-      print(fileName)
-      with open("./csv/" + fileName + ".csv") as csvFile:
+      if os.path.isfile("./img/" + fileName + ".img"):
+        print(fileName + " exists")
+      else:
+        print(fileName)
+        with open("./csv/" + fileName + ".csv") as csvFile:
 
-        csvReader = csv.reader(csvFile, delimiter=",")
-        lineCount = 0
+          csvReader = csv.reader(csvFile, delimiter=",")
+          lineCount = 0
 
-        iterations = []
-        numeriInUso = []
+          iterations = []
+          numeriInUso = []
 
-        for row in csvReader:
-          if lineCount == 0:
-            print("Colum names")
-            lineCount += 1
-          else:
+          for row in csvReader:
+            if lineCount == 0:
+              print("Colum names")
+              lineCount += 1
 
-            iterations.append(row[0])
-            numeriInUso.append(row[1])
-            # print(numeriInUso)
+            else:
+              iterations.append(row[0])
+              numeriInUso.append(row[1])
+                # print(numeriInUso)
+              lineCount += 1
 
-            lineCount += 1
-
-
-      finalFigure = plt.figure()
-      plt.scatter(iterations, numeriInUso)
-      plt.xlabel("Iterazioni")
-      plt.xscale("linear")
-      plt.ylabel("Numero")
-      plt.yscale("linear")
-      # plt.show()
-      finalFigure.savefig("./img/" + fileName + ".png")
+        finalFigure = plt.figure()
+        plt.scatter(iterations, numeriInUso)
+        plt.xlabel("Iterazioni")
+        plt.xscale("linear")
+        plt.ylabel("Numero")
+        plt.yscale("linear")
+        # plt.show()
+        finalFigure.savefig("./img/" + fileName + ".png")
